Refactor task sorting and comparison logic by introducing TaskListSortHelper, removing obsolete use cases, and updating related components
