+++ lib/domain/task_list_sort_helper.dart
@@ -1,12 +1,13 @@
 import 'package:injectable/injectable.dart';
 import 'package:todoapp/data/model/task.dart';
 
-abstract class TasksComparatorUseCase {
+abstract class TaskListSortHelper {
   bool areThemEqual({required List<Task> oldList, required List<Task> newList});
+  List<Task> sortByCompletedStatus(List<Task> tasks);
 }
 
-@Injectable(as: TasksComparatorUseCase)
-class TasksComparatorUseCaseImpl extends TasksComparatorUseCase {
+@Injectable(as: TaskListSortHelper)
+class TaskListSortHelperImpl extends TaskListSortHelper {
   @override
   bool areThemEqual({
     required List<Task> oldList,
@@ -28,5 +29,22 @@ class TasksComparatorUseCaseImpl extends TasksComparatorUseCase {
       return false;
     }
     return true;
+  }
+
+  @override
+  List<Task> sortByCompletedStatus(List<Task> tasks) {
+    List<Task> tasksToBeSorted = List.from(tasks);
+    tasksToBeSorted.sort((a, b) => _sort(a, b));
+    return tasksToBeSorted;
+  }
+
+  int _sort(Task a, Task b) {
+    if (a.isCompleted == false && b.isCompleted) {
+      return -1;
+    } else if (a.isCompleted && b.isCompleted == false) {
+      return 1;
+    } else {
+      return 0;
+    }
   }
 }
